Started on window polymorphism
