+++ lib/main/instance/instance_list.dart
@@ -5,7 +5,7 @@ import 'package:provider/single_child_widget.dart';
 import 'package:smooth_list_view/smooth_list_view.dart';
 
 import '../settings/settings_model.dart';
-import '../setup/theme_setup.dart';
+import '../../setup/theme_setup.dart';
 import 'instance_manager.dart';
 import 'instance_model.dart';
