+++ lib/top_bar/top_bar_items.dart
@@ -7,9 +7,9 @@ import 'package:menu_bar/src/entry.dart';
 import 'package:provider/provider.dart';
 import 'package:window_manager_plus/window_manager_plus.dart';
 
-import '../instance/instance_model.dart';
+import '../main/instance/instance_model.dart';
 import '../process/process_controller.dart';
-import '../settings/settings_model.dart';
+import '../main/settings/settings_model.dart';
 
 /// The Windows buttons in the top bar of the window. This includes the
 /// minimize, maximize, and close buttons.
